Fix compilation problem on Solaris
git-svn-id: http://root.cern.ch/svn/root/trunk@20688 27541ba8-7e3a-0410-8455-c3a389f83636

+++ proofd/src/XrdProofdProtocol.cxx
@@ -4073,7 +4073,7 @@ int XrdProofdProtocol::Create()
       int setupOK = 0;
 
       XrdOucString pmsg = "child process ";
-      pmsg += getpid();
+      pmsg += (int) getpid();
       MTRACE(FORK, "xpd: ", pmsg.c_str());
 
       // We set to the user environment
